+++ conanfile.txt
@@ -1,5 +1,6 @@
 [requires]
 blaze/3.8
+catch2/3.2.1
 
 [generators]
 CMakeDeps
